+++ NEWS.md
@@ -1,5 +1,13 @@
 # CompMetaboTools 0.1
 
+## CompMetaboTools 0.0.8
+
+- Fix issue in `.group_correlation_matrix` that would not group elements even if
+  their correlation is larger than the threshold in some special cases.
+- Add `featureGroupSpectra` function to extract a spectrum for each feature
+  group which can be either a *pseudo spectrum* or a full scan MS1 spectrum.
+
+
 ## CompMetaboTools 0.0.7
 
 - Add `matchRtMz` function to match features based on retention time and m/z